refactor(search): split synthetic line helpers
Keep the primary search line builder focused on assembling positioned spans while preserving the existing augmentation behavior for hyphenated words, stacked labels, and compact table headers.
