+++ src/core/search/lines.ts
@@ -6,35 +6,16 @@ import {
   shouldInsertSemanticSpace,
 } from '../text/spacing.js';
 import { isRtlDominantPositionedText, textOrder } from '../text/textDirection.js';
-import { unionBoxes } from './boxes.js';
+import { isLikelyCompactTableHeaderRow } from './compactTableHeaders.js';
 import { nfkc } from './compiler.js';
-import type { Box, SearchLine, SearchOwner } from './types.js';
+import { withSyntheticSearchLines } from './syntheticLines.js';
+import type { SearchLine, SearchOwner } from './types.js';
 import { buildVerticalSearchLines } from './verticalLines.js';
 
 const DEFAULT_SPACE_GAP_RATIO = 0.25;
 const FONT_SIZE_FALLBACK_PT = 12;
 const SEARCH_SEGMENT_GAP_RATIO = 1.25;
 const SEARCH_SEGMENT_MIN_GAP_PT = 14;
-const HYPHENATED_SEARCH_LINE_SCAN_LIMIT = 6;
-const HYPHENATED_SEARCH_LINE_MAX_GAP_RATIO = 2.5;
-const HYPHENATED_SEARCH_LINE_MAX_GAP_PT = 24;
-const HYPHENATED_SEARCH_LINE_X_TOLERANCE_PT = 12;
-const STACKED_LABEL_SCAN_LIMIT = 24;
-const STACKED_LABEL_MAX_VERTICAL_GAP_RATIO = 1.6;
-const STACKED_LABEL_MAX_VERTICAL_GAP_PT = 14;
-const STACKED_LABEL_CENTER_TOLERANCE_PT = 12;
-const STACKED_LABEL_MAX_CHARS = 30;
-const STACKED_LABEL_MAX_WORDS = 3;
-const STACKED_LABEL_TEXT_RE = /^[\p{L}\p{N}][\p{L}\p{N}\p{M}\s&.,'’()/-]*$/u;
-const COMPACT_TABLE_HEADER_MIN_SPANS = 3;
-const COMPACT_TABLE_HEADER_MAX_SPANS = 8;
-const COMPACT_TABLE_HEADER_MAX_CHARS = 32;
-const COMPACT_TABLE_HEADER_MAX_WORDS = 4;
-const COMPACT_TABLE_HEADER_MAX_ROW_WIDTH_RATIO = 0.75;
-const COMPACT_TABLE_HEADER_MAX_GAP_RATIO = 2.25;
-const COMPACT_TABLE_HEADER_BASE_MAX_GAP_PT = 28;
-const COMPACT_TABLE_HEADER_MAX_GAP_PT = 72;
-const COMPACT_TABLE_HEADER_TEXT_RE = /^[\p{L}\p{N}][\p{L}\p{N}\p{M}\s&.,'’()/%:+-]*$/u;
 
 export function buildSearchLines(spans: readonly TextSpan[] | undefined, pageWidth: number): SearchLine[] {
   if (!spans || spans.length === 0) return [];
@@ -55,7 +36,7 @@ export function buildSearchLines(spans: readonly TextSpan[] | undefined, pageWid
     const xSorted = [...group].sort((a, b) => a.x - b.x);
     const preserveWideWordSpacing = isLikelyWideWordSpacingRow(xSorted, pageWidth);
     const preserveCjkDisplaySpacing = isLikelyCjkDisplaySpacingRow(xSorted);
-    const preserveCompactTableHeader = isLikelyCompactTableHeaderRow(xSorted, pageWidth);
+    const preserveCompactTableHeader = isLikelyCompactTableHeaderRow(xSorted, pageWidth, FONT_SIZE_FALLBACK_PT);
     const segments: TextSpan[][] = [[xSorted[0]]];
 
     for (let i = 1; i < xSorted.length; i++) {
@@ -99,162 +80,7 @@ export function buildSearchLines(spans: readonly TextSpan[] | undefined, pageWid
     }
   }
   const augmented = [...lines, ...buildVerticalSearchLines(spans)];
-  return withHyphenatedSearchLines(withStackedSearchLines(augmented));
-}
-
-function withHyphenatedSearchLines(lines: readonly SearchLine[]): SearchLine[] {
-  const synthetic: SearchLine[] = [];
-  for (let i = 0; i < lines.length; i++) {
-    const line = lines[i];
-    const lineText = line.text.trimEnd();
-    if (!lineText.endsWith('-')) continue;
-    const lineBox = searchLineBox(line);
-    if (!lineBox) continue;
-
-    for (let j = i + 1; j < lines.length && j <= i + HYPHENATED_SEARCH_LINE_SCAN_LIMIT; j++) {
-      const next = lines[j];
-      const nextText = next.text.trimStart();
-      if (!/^[\p{L}\p{N}]/u.test(nextText)) continue;
-      const nextBox = searchLineBox(next);
-      if (!nextBox) continue;
-      const verticalGap = nextBox.y - (lineBox.y + lineBox.height);
-      if (verticalGap < -1) continue;
-      if (
-        verticalGap > Math.max(lineBox.height * HYPHENATED_SEARCH_LINE_MAX_GAP_RATIO, HYPHENATED_SEARCH_LINE_MAX_GAP_PT)
-      ) {
-        break;
-      }
-      if (Math.abs(nextBox.x - lineBox.x) > HYPHENATED_SEARCH_LINE_X_TOLERANCE_PT) continue;
-
-      const trailingSpaces = line.text.length - lineText.length;
-      const leadingSpaces = next.text.length - nextText.length;
-      synthetic.push({
-        text: `${lineText}${nextText}`,
-        owners: [...line.owners.slice(0, line.owners.length - trailingSpaces), ...next.owners.slice(leadingSpaces)],
-        syntheticHyphenated: true,
-      });
-      break;
-    }
-  }
-  return synthetic.length === 0 ? [...lines] : [...lines, ...synthetic];
-}
-
-function withStackedSearchLines(lines: readonly SearchLine[]): SearchLine[] {
-  const synthetic: SearchLine[] = [];
-  for (let i = 0; i < lines.length; i++) {
-    const top = trimmedSearchLine(lines[i]);
-    if (!isStackedLabelText(top.text)) continue;
-    const topBox = searchLineBox(top);
-    if (!topBox) continue;
-
-    for (let j = i + 1; j < lines.length && j <= i + STACKED_LABEL_SCAN_LIMIT; j++) {
-      const bottom = trimmedSearchLine(lines[j]);
-      if (!isStackedLabelText(bottom.text)) continue;
-      const bottomBox = searchLineBox(bottom);
-      if (!bottomBox) continue;
-
-      const verticalGap = bottomBox.y - (topBox.y + topBox.height);
-      if (verticalGap < -1) continue;
-      if (
-        verticalGap > Math.max(topBox.height * STACKED_LABEL_MAX_VERTICAL_GAP_RATIO, STACKED_LABEL_MAX_VERTICAL_GAP_PT)
-      ) {
-        break;
-      }
-      if (!areStackedLabelBoxesAligned(topBox, bottomBox)) continue;
-
-      synthetic.push({
-        text: `${top.text} ${bottom.text}`,
-        owners: [...top.owners, undefined, ...bottom.owners],
-        syntheticStacked: true,
-      });
-      break;
-    }
-  }
-  return synthetic.length === 0 ? [...lines] : [...lines, ...synthetic];
-}
-
-function searchLineBox(line: SearchLine): Box | undefined {
-  const seen = new Set<SearchOwner>();
-  const boxes: Box[] = [];
-  for (const owner of line.owners) {
-    if (!owner || seen.has(owner)) continue;
-    seen.add(owner);
-    boxes.push(owner);
-  }
-  return boxes.length === 0 ? undefined : unionBoxes(boxes);
-}
-
-function trimmedSearchLine(line: SearchLine): SearchLine {
-  const leadingSpaces = line.text.length - line.text.trimStart().length;
-  const trailingSpaces = line.text.length - line.text.trimEnd().length;
-  const end = line.owners.length - trailingSpaces;
-  return {
-    ...line,
-    text: line.text.trim(),
-    owners: line.owners.slice(leadingSpaces, end),
-  };
-}
-
-function isStackedLabelText(text: string): boolean {
-  if (text.length === 0 || text.length > STACKED_LABEL_MAX_CHARS) return false;
-  if (!/\p{L}/u.test(text)) return false;
-  if (text.split(/\s+/u).length > STACKED_LABEL_MAX_WORDS) return false;
-  return STACKED_LABEL_TEXT_RE.test(text);
-}
-
-function areStackedLabelBoxesAligned(top: Box, bottom: Box): boolean {
-  const topCenter = top.x + top.width / 2;
-  const bottomCenter = bottom.x + bottom.width / 2;
-  const centerTolerance = Math.max(STACKED_LABEL_CENTER_TOLERANCE_PT, Math.min(top.width, bottom.width) * 0.75);
-  if (Math.abs(topCenter - bottomCenter) > centerTolerance) return false;
-
-  const overlap = Math.min(top.x + top.width, bottom.x + bottom.width) - Math.max(top.x, bottom.x);
-  return overlap > 0 || Math.abs(topCenter - bottomCenter) <= STACKED_LABEL_CENTER_TOLERANCE_PT;
-}
-
-function isLikelyCompactTableHeaderRow(spans: readonly TextSpan[], pageWidth: number): boolean {
-  if (pageWidth <= 0) return false;
-  if (spans.length < COMPACT_TABLE_HEADER_MIN_SPANS || spans.length > COMPACT_TABLE_HEADER_MAX_SPANS) return false;
-
-  const first = spans[0];
-  const last = spans.at(-1);
-  if (!first || !last) return false;
-  const rowWidth = last.x + last.width - first.x;
-  if (rowWidth > pageWidth * COMPACT_TABLE_HEADER_MAX_ROW_WIDTH_RATIO) return false;
-
-  let letterLabelCount = 0;
-  for (let i = 0; i < spans.length; i++) {
-    const span = spans[i];
-    const text = span.text.trim();
-    if (!isCompactTableHeaderText(text)) return false;
-    if (/\p{L}/u.test(text)) letterLabelCount++;
-
-    const prev = spans[i - 1];
-    if (prev) {
-      const fontSize = span.fontSize || prev.fontSize || FONT_SIZE_FALLBACK_PT;
-      const gap = span.x - (prev.x + prev.width);
-      const baseMaxGap = Math.max(fontSize * COMPACT_TABLE_HEADER_MAX_GAP_RATIO, COMPACT_TABLE_HEADER_BASE_MAX_GAP_PT);
-      const maxGap = Math.max(fontSize * COMPACT_TABLE_HEADER_MAX_GAP_RATIO, COMPACT_TABLE_HEADER_MAX_GAP_PT);
-      if (gap < 0 || gap > maxGap) {
-        return false;
-      }
-      if (gap > baseMaxGap && (isSingleCjkGlyph(prev.text.trim()) || isSingleCjkGlyph(text))) {
-        return false;
-      }
-    }
-  }
-
-  return letterLabelCount >= 2;
-}
-
-function isSingleCjkGlyph(text: string): boolean {
-  return /^[\u3040-\u30ff\u3400-\u9fff\uf900-\ufaff]$/u.test(text);
-}
-
-function isCompactTableHeaderText(text: string): boolean {
-  if (text.length === 0 || text.length > COMPACT_TABLE_HEADER_MAX_CHARS) return false;
-  if (text.split(/\s+/u).length > COMPACT_TABLE_HEADER_MAX_WORDS) return false;
-  return COMPACT_TABLE_HEADER_TEXT_RE.test(text);
+  return withSyntheticSearchLines(augmented);
 }
 
 export function buildOcrSearchLines(words: readonly OcrWord[] | undefined, normalize: boolean): SearchLine[] {
